WIP: In the middle of `types.jl` split into several files and modules. Before extraction field extraction and field assignment from `parsing.jl`
