+++ src/KML.jl
@@ -2,36 +2,47 @@ module KML
 
 # ─── base deps ────────────────────────────────────────────────────────────────
 using OrderedCollections: OrderedDict
-import XML: XML, read as xmlread, parse as xmlparse, write as xmlwrite, Node, LazyNode, nodetype # xml parsing / writing
-using InteractiveUtils: subtypes    # all subtypes of a type
-using StaticArrays                  # small fixed‑size coordinate vectors
-using Automa, Parsers               # for coordinates parsing
+import XML: XML, read as xmlread, parse as xmlparse, write as xmlwrite, Node, LazyNode, nodetype
+using InteractiveUtils: subtypes
+using StaticArrays
+using Automa, Parsers
 
 # ─── split implementation files ──────────────────────────────────────────────
 include("Coordinates.jl")       # coordinate parsing and string generation
-include("Enums.jl")             # KML enum types (altitudeMode, colorMode, etc.)
-include("types.jl")             # all KML data types & helpers (no GeoInterface)
-include("utils.jl")             # utility functions (e.g., for parsing)
-include("Layers.jl")            # Add this line - layer management functionality
-include("geointerface.jl")      # GeoInterface extensions & pretty printing
+include("Enums.jl")             # KML enum types
+include("types.jl")             # all KML data types (now modularized)
+include("utils.jl")             # utility functions
+include("Layers.jl")            # layer management functionality
+include("geointerface.jl")      # GeoInterface extensions
 include("parsing.jl")           # XML → struct & struct → XML
+include("types_integration.jl") # Integration between types and parsing
 include("tables.jl")            # Tables.jl wrapper for Placemarks
-using .TablesBridge             # re‑export ?
-using .Coordinates: coordinate_string  # Export coordinate_string for XML writing
+using .TablesBridge
+using .Coordinates: coordinate_string, Coord2, Coord3
 using .Layers: list_layers, get_layer_names, get_num_layers
 
-# ─── re‑export public names ──────────────────────────────────────────────────
-export KMLFile, LazyKMLFile, Enums, object    # the "root" objects most users need
-export unwrap_single_part_multigeometry       # utility function
-export PlacemarkTable, list_layers, get_layer_names, get_num_layers  # layers info
-export coordinate_string                      # from Coordinates module
+# Remove the manual _parse_kmlfile assignment since it's now in types_integration.jl
 
-for T in vcat(
-    all_concrete_subtypes(KMLElement),        # concrete types
-    all_abstract_subtypes(Object),            # abstract sub‑hierarchy
-)
-    T === KML.Pair && continue                # skip internal helper
-    @eval export $(Symbol(replace(string(T), "KML." => "")))
+# ─── re‑export public names ──────────────────────────────────────────────────
+export KMLFile, LazyKMLFile, Enums, object
+export unwrap_single_part_multigeometry
+export PlacemarkTable, list_layers, get_layer_names, get_num_layers
+export coordinate_string, Coord2, Coord3
+
+# Export all types from types.jl (which re-exports from submodules)
+for name in names(Core; all=false)
+    if name != :Core && name != :eval && name != :include
+        @eval export $name
+    end
+end
+
+# Also export from all the submodules loaded by types.jl
+for mod in [TimeElements, Components, Styles, Views, Geometries, Features]
+    for name in names(mod; all=false)
+        if name != nameof(mod)
+            @eval export $name
+        end
+    end
 end
 
 function __init__()
@@ -40,6 +51,3 @@ function __init__()
 end
 
 end # module
-
-# ─── Issues ──────────────────────────────────────────────────
-# - [ ] `list_layers`does not report the correct number of placemarks in each layer of "NEON_Field_Sites_KMZ_v20_May2025.kmz"
